refactored the mainController and took out all functions that have to do with notes
